initial migration of streamlined MATLAB VEFIE code
the frequency dependency for the materials is implemented so it won't match the original results exactly but you can see it is close. the issue here is that when the geometry is changed to the landscape, no calculation takes place. so the code is broken somehow. this will now be investigated and fixed.